Add extended SPI HAL module driver implementation
This commit introduces the stm32g0xx_hal_spi_ex.c file, which provides extended functionalities for the SPI peripheral. The new implementation includes an IO operation function to flush the RX FIFO, enhancing the SPI driver capabilities. The HAL_SPIEx_FlushRxFifo function allows for efficient management of received data by clearing the RX FIFO until it is empty or a timeout occurs. This addition is part of the ongoing effort to improve the STM32G0xx HAL driver.